pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "amigo"
version = "0.1.0"
description = "AI companion agent backend built on LiveKit Agents"
requires-python = ">=3.10, <3.14"

dependencies = [
    "httpx>=0.28.1",
    "livekit-agents[deepgram,elevenlabs,google,silero,turn-detector]~=1.3",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dateutil",
    "python-dotenv",
    "redis[hiredis]>=5.0.0",
    "typer>=0.19.2",
]

[dependency-groups]
dev = [
    "fakeredis[lua]",
    "pytest",
    "pytest-asyncio",
    "ruff",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "B", "A", "C4", "UP", "SIM", "RUF"]
ignore = ["E501", "E402"]  # Line too long; module-level import not at top (needed for env setup)

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
